add: при изменение имени канала, он меняется при помощи сокета
